requirements.txt
openai
python-dotenv
requests
tqdm
pydantic
duckduckgo-search
ddgs
